chore(engine): Split response serialization code
Intend to touch that code later, so splitting it up in this PR to
improve the readability.
